requirements.txt
 python-telegram-bot[job-queue]==20.7
 fastapi==0.116.1
 uvicorn==0.35.0
 sqlalchemy==2.0.30
 alembic==1.16.4
 asyncpg==0.30.0
 python-dotenv==1.0.0
 yookassa==3.6.0
 httpx==0.25.2
 aiohttp==3.12.15
 pydantic==2.11.7
 requests==2.32.4
 aiofiles==23.2.0